refactor(Models): Models 폴더 추가 및 클래스 경로 수정
Models 폴더에 데이터를 담은 클래스 파일 저장
이동에 따른 import문 경로 수정
